[pod] Add tips for installing "Headless" and on Debian/Ubuntu
The package list was what I found I needed when trying to install on a mostly
vanilla headless Ubutnu 12.04

The xvfb-run was what potyl suggested when I ran into problems.

It ought to be kinda moot now, due to inclusion of Test::NeedsDisplay in the
Makefile.PL and test files.